requirements.txt
ezdxf>=1.1.0
shapely>=2.0.0
PySide6>=6.6.0
chromadb>=0.4.0
numpy>=1.24.0
google-generativeai>=0.3.0
rtree>=1.0.0
